Refactor: Pass emote codes to a single component, instead of having many.
