+++ src/components/Ava/Animated.js
@@ -1,18 +1,20 @@
-import React, { useCallback, useEffect, useState } from 'react';
+import React, { useEffect, useState } from 'react';
 import PropTypes from 'prop-types';
 import { countBy } from 'lodash';
 
+import styled from 'styled-components';
 import { motion } from 'framer-motion';
 
 import EmoteBucketSystem from 'helpers/EmoteBucketSystem';
 import { useChatContext, useImageryContext } from 'providers';
 
-import * as Poses from './Poses';
+import Pose from './Pose';
+import { url, path } from './constants';
 
 function Animated({ className }) {
   const { connected, client } = useChatContext();
   const { emotes } = useImageryContext();
-  const [emoteCodes, setEmoteCodes] = useState();
+  const [emoteCodes, setEmoteCodes] = useState([]);
   const [currentPose, setCurrentPose] = useState('avalonBASE');
 
   const ebs = new EmoteBucketSystem({
@@ -24,35 +26,23 @@ function Animated({ className }) {
     }
   });
 
-  const processEmotes = useCallback(
-    message => {
-      const occurances = countBy(message.split(' '), token => token);
-      const filtered = Object.keys(occurances)
-        .filter(key => emoteCodes.includes(key))
-        .reduce((obj, key) => {
-          // eslint-disable-next-line no-param-reassign
-          obj[key] = occurances[key];
-          return obj;
-        }, {});
-      return Object.keys(filtered).forEach(key =>
-        ebs.processIncomgingEmote(key)
-      );
-    },
-    [emoteCodes, ebs]
-  );
+  const processEmotes = message => {
+    const occurances = countBy(message.split(' '), token => token);
+    const filtered = Object.keys(occurances)
+      .filter(key => emoteCodes.includes(key))
+      .reduce((obj, key) => {
+        // eslint-disable-next-line no-param-reassign
+        obj[key] = occurances[key];
+        return obj;
+      }, {});
+    return Object.keys(filtered).forEach(key => {
+      ebs.processIncomgingEmote(key, occurances[key]);
+    });
+  };
 
   const resetPose = () => {
     setCurrentPose('avalonBASE');
   };
-
-  const availablePoses = () => ({
-    avalonBASE: <Poses.avalonBASE />,
-    avalonHEHE: <Poses.avalonHEHE callback={resetPose} />
-  });
-
-  useEffect(() => {
-    console.log('currentPose', currentPose);
-  }, [currentPose]);
 
   useEffect(() => {
     const codes = emotes.map(emote => emote.code);
@@ -67,13 +57,17 @@ function Animated({ className }) {
   }, [connected, client, processEmotes]);
 
   return (
-    <motion.div
+    <Wrapper
       className={className}
       initial={{ opacity: 0 }}
       animate={{ opacity: 1 }}
     >
-      {availablePoses()[currentPose]}
-    </motion.div>
+      {currentPose === 'avalonBASE' ? (
+        <img src={`${url}${path}/avalonBASE.png`} alt="avalonBASE" />
+      ) : (
+        <Pose name={currentPose} callback={resetPose} />
+      )}
+    </Wrapper>
   );
 }
 
@@ -84,5 +78,11 @@ Animated.propTypes = {
 Animated.defaultProps = {
   className: ''
 };
+
+const Wrapper = styled(motion.div)`
+  img {
+    display: block;
+  }
+`;
 
 export default Animated;
