--- src/components/Ava/Poses/avalonBASE.js
@@ -1,9 +1,0 @@
-import React from 'react';
-
-import { url, path } from './constants';
-
-function Pose() {
-  return <img src={`${url}${path}/avalonBASE.png`} alt="avalonBASE" />;
-}
-
-export default Pose;
