--- src/components/Ava/Poses/index.js
@@ -1,2 +1,0 @@
-export { default as avalonBASE } from './avalonBASE';
-export { default as avalonHEHE } from './avalonHEHE';
