+++ src/components/Ava/poses.js
@@ -1,0 +1,3 @@
+export default {
+  avalonHEHE: { steps: 51 }
+};
